Update Skip Types to make it easier to directly embed files into prompts from deployed package inside node_modules in target environments (or from within MJ repo via TS file)
