+++ packages/SkipTypes/src/component-types.ts
@@ -1,413 +1,8 @@
 /**
- * @fileoverview Component interface types for Skip API
- * 
- * This file contains types that define the interface between Skip-generated components
- * and their container applications. These types enable rich, interactive components that
- * can communicate with the host MemberJunction environment. The types define:
- * 
- * - Component callback interfaces for parent-child communication (SkipComponentCallbacks)
- * - Component object interfaces that Skip components must implement (SkipComponentObject)
- * - Initialization parameters and utility interfaces (SkipComponentInitParams, SkipComponentUtilities)
- * - Styling and theming interfaces (SkipComponentStyles)
- * - Simplified metadata and data access interfaces (SimpleMetadata, SimpleRunView, SimpleRunQuery)
- * 
- * Components generated by Skip implement the SkipComponentObject interface and expose
- * themselves globally on the window object, allowing container applications to initialize,
- * refresh, and print them. The components can call back to the container through the provided
- * callback functions to refresh data, open records, update user state, and send custom events.
- * 
- * The utility interfaces provide Skip components with controlled access to the MemberJunction
- * system for dynamic data access, metadata queries, and view execution without requiring
- * full access to the underlying MJ APIs.
- * 
- * @author MemberJunction
- * @since 2.0.0
+ * This file is often **DIRECTLY** embedded into prompts so keep it clean and simple and limited to the types below.
  */
 
-import { CompositeKey } from "@memberjunction/core";
-import { SimpleDataContext } from "./shared";
-import { SimpleMetadata, SimpleRunQuery, SimpleRunView } from "./shared";
 import { DataContext } from "@memberjunction/data-context";
-
-/**
- * This interface defines the available callback functions that a Skip component might call in the parent.
- */
-export interface SkipComponentCallbacks {
-    /**
-     * The component can invoke this method in the callbacks object, when provided, to refresh the data context 
-     * and that will in turn result in the component's init function being called again with the new data context.
-     * @returns 
-     */
-    RefreshData: () => void;
-
-    /**
-     * If an action occurs inside a component where it would be desirable for the containing UI to open a specific 
-     * record, if supported, this event can be listened to and the container UI can then open the record.
-     * @param entityName - this is the Entity NAME from the Entity metadata, not the table name or base view name. Use Entity Metadata to provide the entity name here
-     * @param key - this is an array of key/value pairs representing the primary key. The format of a Composite Key is an array of KeyValuePair objects and KeyValuePair objects simply have FieldName and Value properties. In most cases entities have single-valued primary keys but this structure is here for complex entity types that have composite primary keys
-     * @returns 
-     */
-    OpenEntityRecord: (entityName: string, key: CompositeKey) => void;
-
-    /**
-     * This event should be raised by the HTML component whenever something changes within the component that should be tracked as a change in state
-     * that will persist. userState is any valid, simple JavaScript object, meaning it can have scalars, arrays, objects, etc, it must be an object that 
-     * can be serialized to JSON, but otherwise has no special requirements. The parent component will be responsible for tracking the user-specific states
-     * and passing them back to the HTML component each time it is loaded or if the user changes via the init function.
-     * @param userState 
-     * @returns 
-     */
-    UpdateUserState: (userState: any) => void;
-
-    /**
-     * Used for any other type of event notification that a component might want to send to the parent component.
-     * @param eventName 
-     * @param eventData 
-     * @returns 
-     */
-    NotifyEvent: (eventName: string, eventData: any) => void;
-}
-
-/**
- * This is the function signature for the initialization function provided by each Skip component via the SkipComponentObject so that a container can interact with it.
- * This function is called when the component is loaded by its container. The function receives the data context, an optional userState property, and a set of callbacks that can be used to interact with the parent component.
- * userState is an optional parameter that can be used to pass in any state information that the parent component wants to provide to the component that is specific
- * to the CURRENT user. If the component modifies the userState, it should notify the parent component via the UserStateChanged event in the callbacks object so that the parent component can handle storage.
- */
-export type SkipComponentInitFunction = (params: SkipComponentInitParams) => void;
-
-/**
- * This is the function signature for the print function that is provided by the component via the SkipComponentObject
- */
-export type SkipComponentPrintFunction = () => void;
-/**
- * This is the function signature for the refresh function that is provided by the component via the SkipComponentObject
- */
-export type SkipComponentRefreshFunction = () => void;
-
-/**
- * Parameters that are passed to the SkipComponentInitFunction when it is called by the parent component.
- */
-export interface SkipComponentInitParams {
-    /**
-     * Contains the static data specified by the root component specification. This data is pre-loaded and passed to 
-     * the component during initialization and anytime the component is refreshed.
-     */
-    staticData: SimpleDataContext;
-    /**
-     * Contains the dynamic data utilities that the component can use to access MemberJunction data. This includes
-     * the metadata, run view, and run query utilities that the component can use to access data dynamically at runtime.
-     * The component can use these utilities to fetch data at runtime based on the user's interactions with the component.
-     */
-    utilities?: SkipComponentUtilities;
-    /**
-     * Any valid serializable JavaScript object that represents the user-specific state for the component.
-     */
-    userState?: any;
-    /**
-     * Callbacks that the component can use to interact with the parent component. These callbacks allow the component to refresh data, open records, update user state, and send custom events.
-     */
-    callbacks?: SkipComponentCallbacks;
-    /**
-     * The default styles the component should use as specified by its parent (the container in the case of a root component, or a parent component in the case 
-     * of a child component). The component can alter these styles based on the requirements specified by the user in its design documentation. 
-     */
-    styles?: SkipComponentStyles;
-}
-
-/**
- * This interface defines styles that can be applied to the component. The container can provide
- * styles to the top level component. The top level component can alter these styles based on
- * the prompting of the user, learned notes, etc, and adjust the styles of the component accordingly. In addition
- * the top level component will pass in its computed styles to each sub-component so that the sub-components
- * can do the same recursively down to any level of depth. This allows sub-components to inherit styles but
- * also make adjustments as required based on functional needs and user input.
- */
-export interface SkipComponentStyles {
-    colors: {
-        // Primary palette
-        primary: string // '#5B4FE9',
-        primaryHover: string // '#4940D4',
-        primaryLight?: string // '#E8E6FF',
-        
-        // Secondary palette
-        secondary: string // '#64748B',
-        secondaryHover?: string // '#475569',
-        
-        // Status colors
-        success: string // '#10B981',
-        successLight?: string // '#D1FAE5',
-        warning?: string // '#F59E0B',
-        warningLight?: string // '#FEF3C7',
-        error?: string // '#EF4444',
-        errorLight?: string // '#FEE2E2',
-        info?: string // '#3B82F6',
-        infoLight?: string // '#DBEAFE',
-        
-        // Base colors
-        background: string // '#FFFFFF',
-        surface: string // '#F8FAFC',
-        surfaceHover?: string // '#F1F5F9',
-        
-        // Text colors
-        text: string // '#1E293B',
-        textSecondary: string // '#64748B',
-        textTertiary?: string // '#94A3B8',
-        textInverse?: string // '#FFFFFF',
-        
-        // Border colors
-        border: string // '#E2E8F0',
-        borderLight?: string // '#F1F5F9',
-        borderFocus?: string // '#5B4FE9',
-        
-        // Shadow colors
-        shadow?: string // 'rgba(0, 0, 0, 0.05)',
-        shadowMedium?: string // 'rgba(0, 0, 0, 0.1)',
-        shadowLarge?: string // 'rgba(0, 0, 0, 0.15)',
-        
-        // Allow additional custom colors
-        [key: string]: string | undefined;
-    };
-    spacing: {
-        xs: string // '4px',
-        sm: string // '8px',
-        md: string // '16px',
-        lg: string // '24px',
-        xl: string // '32px',
-        xxl?: string // '48px',
-        xxxl?: string // '64px',
-        
-        // Allow additional custom spacing
-        [key: string]: string | undefined;
-    };
-    typography: {
-        fontFamily: string //'-apple-system, BlinkMacSystemFont, "Inter", "Segoe UI", Roboto, sans-serif',
-        fontSize: {
-          xs?: string // '12px',
-          sm: string // '14px',
-          md: string // '16px',
-          lg: string // '18px',
-          xl: string // '24px',
-          xxl?: string // '32px',
-          xxxl?: string // '40px',
-          
-          // Allow additional custom sizes
-          [key: string]: string | undefined;
-        },
-        fontWeight?: {
-          light?: string // '300',
-          regular?: string // '400',
-          medium?: string // '500',
-          semibold?: string // '600',
-          bold?: string // '700',
-          
-          // Allow additional custom weights
-          [key: string]: string | undefined;
-        },
-        lineHeight?: {
-          tight?: string // '1.25',
-          normal?: string // '1.5',
-          relaxed?: string // '1.75',
-          
-          // Allow additional custom line heights
-          [key: string]: string | undefined;
-        }
-    };
-    borders: {
-        radius: string | { // Allow either a single string or an object with multiple radius options
-          sm?: string // '6px',
-          md?: string // '8px',
-          lg?: string // '12px',
-          xl?: string // '16px',
-          full?: string // '9999px',
-          
-          // Allow additional custom radii
-          [key: string]: string | undefined;
-        };
-        width: string | { // Allow either a single string or an object with multiple width options
-          thin?: string // '1px',
-          medium?: string // '2px',
-          thick?: string // '3px',
-          
-          // Allow additional custom widths
-          [key: string]: string | undefined;
-        };
-    }
-    shadows?: {
-        sm?: string // '0 1px 2px 0 rgba(0, 0, 0, 0.05)',
-        md?: string // '0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06)',
-        lg?: string // '0 10px 15px -3px rgba(0, 0, 0, 0.1), 0 4px 6px -2px rgba(0, 0, 0, 0.05)',
-        xl?: string // '0 20px 25px -5px rgba(0, 0, 0, 0.1), 0 10px 10px -5px rgba(0, 0, 0, 0.04)',
-        inner?: string // 'inset 0 2px 4px 0 rgba(0, 0, 0, 0.06)',
-        
-        // Allow additional custom shadows
-        [key: string]: string | undefined;
-    }
-    transitions?: {
-        fast?: string // '150ms ease-in-out',
-        normal?: string // '250ms ease-in-out',
-        slow?: string // '350ms ease-in-out',
-        
-        // Allow additional custom transitions
-        [key: string]: string | undefined;
-    }
-    overflow: string
-}
-
-/**
- * This is the interface that each Skip component will expose to the parent component and assign it a name globally on the window object so that the parent component can call it.
- * The component will create this object and it will include the members defined in this interface.
- */
-export interface SkipComponentObject {
-    /**
-     * The React component that Angular will render directly using ReactDOM.
-     * This component receives props including data, userState, callbacks, utilities, and styles.
-     */
-    component: any; // really a React.ComponentType<RootComponentProps>;
-
-    /**
-     * The optional print function that is called when the user clicks on the print button in the parent of the component. This function will never be called by the parent before the init function so the print function
-     * can assume the component has been initialized;
-     */
-    print?: SkipComponentPrintFunction;
-
-    /**
-     * The optional refresh function that is called when the user clicks on the refresh button in the parent of the component. This function will never be called by the parent before the init function so the refresh function
-     */
-    refresh?: SkipComponentRefreshFunction;
-}
-
-/**
- * This interface defines the utilities that are available to the Skip component. These utilities are used to interact with the host MemberJunction system to 
- * retrieve metadata, run views, and run queries. The utilities are passed into the SkipComponentInitFunction by the container.
- */
-export interface SkipComponentUtilities {
-    md: SimpleMetadata,
-    rv: SimpleRunView,
-    rq: SimpleRunQuery
-}
-
-/**
- * Defines a given option for a generated component that the user can choose. The code/componentObjectName properties are used to render the component in the UI.
- */
-export type SkipComponentOption = {
-    /**
-     * Full details of the generated component option including functional, technical, code, and child componentry.
-     */
-    option: SkipComponentRootSpec;
-
-    /**
-     * If multiple component options are provided for a given @interface SkipAPIAnalysisCompleteResponse, a "judge" AI will evaluate all the functional
-     * responses and will rank order them with an explanation of why they were each ranked that way. Rankings are not absolute, they are relative to the
-     * # of components contained within an array of SkipComponentOption types.  
-     */
-    AIRank: number | undefined;
-    /**
-     * The AI's explanation of why it ranked the component the way it did. This is useful for understanding the AI's reasoning and can be used to improve future components 
-     * as well as provide context to the user about why a particular component was chosen as the best option.
-     */
-    AIRankExplanation: string | undefined;
-    /**
-     * The user's provided feedback on the component option. Unlike the AIRank, this is a subjective rating provided by the user and is 
-     * a number between 1 and 10, where 1 is the lowest rating and 10 is the highest rating.
-     */
-    UserRank: number | undefined;
-    /**
-     * If the host application provides a way for the user to provide feedback on the component option, 
-     * this is the explanation of why the user rated the component the way they did if they provided feedback.
-     */
-    UserRankExplanation: string | undefined;
-}
-
-/**
- * Defines the data requirements for a Skip component, supporting three different data access modes.
- * This interface is critical for understanding how components interact with MemberJunction data
- * and helps optimize performance by clearly defining data access patterns.
- * 
- * @since 2.1.0 - Enhanced with comprehensive support for static, dynamic, and hybrid data modes
- */
-export interface SkipComponentDataRequirements {
-    /**
-     * The primary data access mode for this component.
-     * - 'static': Data is pre-loaded and passed to the component during initialization
-     * - 'dynamic': Component fetches data at runtime using MJ utilities
-     * - 'hybrid': Component uses both static and dynamic data access patterns
-     */
-    mode: 'static' | 'dynamic' | 'hybrid';
-    
-    /**
-     * For static mode: References to data context items that this component uses.
-     * These are pre-loaded data snapshots that are passed to the component during initialization.
-     * Static mode is preferred for:
-     * - Reports with fixed datasets
-     * - Components where users don't need entity-level permissions
-     * - Scenarios requiring reduced database load
-     */
-    staticData?: {
-        /**
-         * Reference to the data context that this component uses.
-         * Points to a DataContext object containing pre-loaded data.
-         */
-        dataContext: DataContext;
-        
-        /**
-         * Optional description of how the static data is used by the component
-         */
-        description?: string;
-    };
-    
-    /**
-     * For dynamic mode: Defines which MemberJunction entities this component needs access to.
-     * The component will use the RunView/RunQuery utilities to fetch data at runtime.
-     * Dynamic mode is preferred for:
-     * - Interactive dashboards with drill-down capabilities
-     * - Components requiring real-time data
-     * - Scenarios where users need entity-level permission validation
-     */
-    dynamicData?: {
-        /**
-         * Array of entity names that this component needs to access dynamically.
-         * These should be valid MemberJunction entity names.
-         */
-        requiredEntities: string[];
-        
-        /**
-         * Optional description of the dynamic data access patterns
-         */
-        description?: string;
-    };
-    
-    /**
-     * For hybrid mode: Components can use both static and dynamic data access.
-     * Both staticData and dynamicData sections should be populated.
-     * Hybrid mode is useful for:
-     * - Components with both summary (static) and detail (dynamic) views
-     * - Optimizing performance while maintaining interactivity
-     */
-    hybridStrategy?: {
-        /**
-         * Description of how the component decides when to use static vs dynamic data
-         */
-        description: string;
-        
-        /**
-         * Optional performance considerations for the hybrid approach
-         */
-        performanceNotes?: string;
-    };
-    
-    /**
-     * General description of the component's data requirements and access patterns.
-     * This should provide a high-level overview of the data strategy.
-     */
-    description?: string;
-    
-    /**
-     * Security considerations for data access
-     * @since 2.1.0
-     */
-    securityNotes?: string;
-}
 
 /**
  * Represents a complete specification for a generated Skip component, including its structure,
@@ -579,4 +174,131 @@ export interface SkipComponentChildSpec {
      * An array of sub-components (recursive structure)
      */
     components: SkipComponentChildSpec[];
+}
+
+
+/**
+ * Defines the data requirements for a Skip component, supporting three different data access modes.
+ * This interface is critical for understanding how components interact with MemberJunction data
+ * and helps optimize performance by clearly defining data access patterns.
+ * 
+ * @since 2.1.0 - Enhanced with comprehensive support for static, dynamic, and hybrid data modes
+ */
+export interface SkipComponentDataRequirements {
+    /**
+     * The primary data access mode for this component.
+     * - 'static': Data is pre-loaded and passed to the component during initialization
+     * - 'dynamic': Component fetches data at runtime using MJ utilities
+     * - 'hybrid': Component uses both static and dynamic data access patterns
+     */
+    mode: 'static' | 'dynamic' | 'hybrid';
+    
+    /**
+     * For static mode: References to data context items that this component uses.
+     * These are pre-loaded data snapshots that are passed to the component during initialization.
+     * Static mode is preferred for:
+     * - Reports with fixed datasets
+     * - Components where users don't need entity-level permissions
+     * - Scenarios requiring reduced database load
+     */
+    staticData?: {
+        /**
+         * Reference to the data context that this component uses.
+         * Points to a DataContext object containing pre-loaded data.
+         */
+        dataContext: DataContext;
+        
+        /**
+         * Optional description of how the static data is used by the component
+         */
+        description?: string;
+    };
+    
+    /**
+     * For dynamic mode: Defines which MemberJunction entities this component needs access to.
+     * The component will use the RunView/RunQuery utilities to fetch data at runtime.
+     * Dynamic mode is preferred for:
+     * - Interactive dashboards with drill-down capabilities
+     * - Components requiring real-time data
+     * - Scenarios where users need entity-level permission validation
+     */
+    dynamicData?: {
+        /**
+         * Array of objects specifying the exact entities and fields the component will 
+         * need to fulfill user requirements. This detail also allows other components
+         * to not have to know the database schema.
+         */
+        requiredEntities: SkipComponentEntityDataRequirement[];
+        
+        /**
+         * Optional description of the dynamic data access patterns
+         */
+        description?: string;
+    };
+    
+    /**
+     * For hybrid mode: Components can use both static and dynamic data access.
+     * Both staticData and dynamicData sections should be populated.
+     * Hybrid mode is useful for:
+     * - Components with both summary (static) and detail (dynamic) views
+     * - Optimizing performance while maintaining interactivity
+     */
+    hybridStrategy?: {
+        /**
+         * Description of how the component decides when to use static vs dynamic data
+         */
+        description: string;
+        
+        /**
+         * Optional performance considerations for the hybrid approach
+         */
+        performanceNotes?: string;
+    };
+    
+    /**
+     * General description of the component's data requirements and access patterns.
+     * This should provide a high-level overview of the data strategy.
+     */
+    description?: string;
+    
+    /**
+     * Security considerations for data access
+     * @since 2.1.0
+     */
+    securityNotes?: string;
+}
+
+/**
+ * Specifications for the use of a single entity within a Skip component.
+ */
+export type SkipComponentEntityDataRequirement = {
+    /**
+     * The name of the entity as defined in the metadata.
+     */
+    entityName: string;
+
+    /**
+     * Description of the entityt as defined in the metadata.
+     */
+    entityDescription?: string;
+
+    /**
+     * These fields can be shown to the user as determined by each component.
+     */
+    displayFields: string[];
+
+    /**
+     * These fields can be used when filtering data as appropriately determined by each component.
+     */
+    filterFields?: string[];
+
+    /**
+     * These fields can be used when sorting data as appropriately determined by each component.
+     */
+    sortFields?: string[];
+
+    /**
+     * Optional, description of when/where/how components should use this entity.
+     */
+    usageContext?: string;
 }
